--- app/helpers/response_service.rb
@@ -1,3 +1,0 @@
-class ResponseService
-  
-end
