+++ app/models/response.rb
@@ -6,7 +6,7 @@ class Response < ApplicationRecord
 
   belongs_to :response_map, class_name: 'ResponseMap', foreign_key: 'map_id', inverse_of: false
   has_many :scores, class_name: 'Answer', foreign_key: 'response_id', dependent: :destroy, inverse_of: false
-
+  
   alias map response_map
   delegate :questionnaire, :reviewee, :reviewer, to: :map
   
@@ -16,7 +16,8 @@ class Response < ApplicationRecord
   # e.g. student click "Begin" or "Update" to start filling out a rubric for others' work
   # if false: we figure out which questionnaire to display base on @response object
   # e.g. student click "Edit" or "View"
-  def set_content(new_response = false)
+  def set_content(new_response = false, action)
+    assign_action_parameters
     @title = @map.get_title
     if @map.survey?
       @survey_parent = @map.survey_parent
@@ -59,59 +60,5 @@ class Response < ApplicationRecord
     end
     @return = params[:return]
   end
-  def reportable_difference?
-    map_class = map.class
-    # gets all responses made by a reviewee
-    existing_responses = map_class.assessments_for(map.reviewee)
-
-    count = 0
-    total = 0
-    # gets the sum total percentage scores of all responses that are not this response
-    existing_responses.each do |response|
-      unless id == response.id # the current_response is also in existing_responses array
-        count += 1
-        total +=  response.aggregate_questionnaire_score.to_f / response.maximum_score
-      end
-    end
-
-    # if this response is the only response by the reviewee, there's no grade conflict
-    return false if count.zero?
-
-    # calculates the average score of all other responses
-    average_score = total / count
-
-    # This score has already skipped the unfilled scorable question(s)
-    score = aggregate_questionnaire_score.to_f / maximum_score
-    questionnaire = questionnaire_by_answer(scores.first)
-    assignment = map.assignment
-    assignment_questionnaire = AssignmentQuestionnaire.find_by(assignment_id: assignment.id, questionnaire_id: questionnaire.id)
-
-    # notification_limit can be specified on 'Rubrics' tab on assignment edit page.
-    allowed_difference_percentage = assignment_questionnaire.notification_limit.to_f
-
-    # the range of average_score_on_same_artifact_from_others and score is [0,1]
-    # the range of allowed_difference_percentage is [0, 100]
-    (average_score - score).abs * 100 > allowed_difference_percentage
-  end
-
-  def aggregate_questionnaire_score
-    # only count the scorable questions, only when the answer is not nil
-    # we accept nil as answer for scorable questions, and they will not be counted towards the total score
-    sum = 0
-    scores.each do |s|
-      question = Question.find(s.question_id)
-      # For quiz responses, the weights will be 1 or 0, depending on if correct
-      sum += s.answer * question.weight unless s.answer.nil? || !question.scorable?
-    end
-    sum
-  end
-  def sort_review_scores
-    @sorted = @review_scores.sort do |m1, m2|
-      if m1.version_num.to_i && m2.version_num.to_i
-        m2.version_num.to_i <=> m1.version_num.to_i
-      else
-        m1.version_num ? -1 : 1
-      end
-    end
-  end
+  
 end
